+++ RELEASE_NOTES.md
@@ -1,5 +1,5 @@
 ### 1.1.0
-* WIP
+* Added support for a nested activator in `PerControllerConfigActivator`
 
 ### 1.0 - 23 March 2015
 * First release.